services/api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.2.1
redis==5.0.4
SQLAlchemy==2.0.31
asyncpg==0.29.0
httpx==0.27.0
pyyaml==6.0.2
binance-connector==3.6.0
websockets==12.0